ChatBotStudent/backend/requirements.txt
# Django Framework
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0

# AI/ML Libraries - Core (FIXED VERSIONS)
transformers==4.35.2
torch==2.3.1
sentence-transformers==2.7.0
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.2.2

# Vector Search & FAISS
faiss-cpu==1.7.4

# ✅ Speech-to-Text (Whisper)
faster-whisper>=0.10.0

# ✅ REQUIRED: Text-to-Speech Dependencies
gTTS==2.4.0
pydub==0.25.1

# ✅ Audio Processing Dependencies (COMPATIBLE VERSIONS)
av>=10.0.0
tokenizers>=0.14.0
huggingface-hub==0.19.4

# ✅ OPTIONAL: Enhanced Audio Processing
# Uncomment if you need advanced audio features
# librosa>=0.10.0
# soundfile>=0.12.0

# ✅ REQUIRED: Google AI Integration
google-generativeai==0.3.2

# ✅ REQUIRED: LangChain for Agent System
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.20
langchain-core>=0.1.0

# HTTP Requests (for APIs)
requests==2.31.0
urllib3==2.1.0

# ✅ REQUIRED: Authentication & Security  
PyJWT==2.8.0
bcrypt==4.1.2

# ✅ REQUIRED: Database Support
psycopg2-binary==2.9.9  # PostgreSQL support
dj-database-url==1.2.0  # Database URL parsing
# sqlite3 is built-in with Python

# ✅ REQUIRED: Vietnamese Language Processing
underthesea==6.7.0
pyvi==0.1.1

# ✅ REQUIRED: File Processing
openpyxl==3.1.2
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
Pillow==9.5.0

# ✅ Logging & Monitoring
colorama>=0.4.6

# ✅ Unicode & Text Processing
unidecode>=1.3.6
regex>=2023.0.0

# ✅ REQUIRED: Google Services Integration
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# ✅ REQUIRED: Development Tools
python-decouple==3.8

# ✅ REQUIRED: PDF & DOCX Processing (OCR)
pdf2image==1.17.0
pytesseract==0.3.10